add DatePicker change money from Int to Double

+++ app/src/main/java/com/example/myfinance/feature/domain/model/Account.kt
@@ -2,6 +2,6 @@ package com.example.myfinance.feature.domain.model
 
 data class Account(
     val id: Int,
-    val balance: Int,
+    val balance: Double,
     val currency: String = "₽"
 )
